+++ src/Utils.php
@@ -3,7 +3,6 @@
 namespace Gaambo\DeployerUtils;
 
 use function Deployer\output;
-use function Deployer\quote as deployerQuote;
 
 class Utils
 {
@@ -23,27 +22,5 @@ class Utils
         }
 
         return $verbosityArgument;
-    }
-
-    /**
-     * @param array<mixed> $array
-     * @return array<string>
-     */
-    public static function parseStringArray(array $array): array
-    {
-        return array_values(array_filter(array_map(
-            fn($value) => is_string($value) ? $value : null,
-            $array
-        )));
-    }
-
-    public static function parseStringOrNull(mixed $string): ?string
-    {
-        return is_string($string) ? $string : null;
-    }
-
-    public static function quote(string $arg): string
-    {
-        return deployerQuote($arg);
     }
 }